Move db-related files to the helm plugin
